[#367] Rename module
