+++ frontend/src/modules/signup-old/view-style.scss
